# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nutrio food discovery and enquiry API
servers:
  - url: /api
    description: Base API path
...
paths:
  /openai/assistant:
    get:
      operationId: getAssistantHistory
      responses:
        "200":
          description: Private saved chat and remaining daily allowance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssistantHistory" }
        "401": { description: Sign in required }
    delete:
      operationId: deleteAssistantHistory
      responses:
        "204": { description: Saved chat removed; daily usage allowance is not reset }
        "401": { description: Sign in required }
        "409": { description: Wait for the current reply before clearing }
  /openai/assistant/messages:
    post:
      operationId: sendAssistantMessage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssistantInput" }
      responses:
        "200":
          description: Saved reply, or the state of an existing request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssistantTurn" }
        "400": { description: Invalid question }
        "401": { description: Sign in required }
        "409": { description: Another question is being answered or retry ID conflicts }
        "429": { description: Daily AI allowance reached }
        "503": { description: Assistant unavailable; question preserved for retry }
...
  /member/challenges:
    get:
      operationId: getMemberChallenges
      responses:
        "200":
          description: Own joined and previously left challenges
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChallengeProgress" }
        "401": { description: Sign in required }
  /member/challenges/{slug}/membership:
    put:
      operationId: setChallengeMembership
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChallengeMembershipInput" }
      responses:
        "200":
          description: Joined, resumed or left; existing progress retained
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChallengeProgress" }
        "401": { description: Sign in required }
        "404": { description: Challenge or membership not found }
  /member/challenges/{slug}/check-in:
    put:
      operationId: setChallengeCheckin
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChallengeCheckinInput" }
      responses:
        "200":
          description: Today's check-in or undo saved; repeated requests are idempotent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChallengeProgress" }
        "401": { description: Sign in required }
        "404": { description: Membership not found }
        "409": { description: Join an active challenge within its window first }
...
  /articles/{slug}:
    get:
      operationId: getArticle
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: A published article
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Article" }
        "404":
          description: Article not found
  /newsletter/signups:
    post:
      operationId: createNewsletterSignup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSignupInput"
      responses:
        "200":
          description: Signup recorded, including duplicate-safe submissions
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { type: string }
        "400":
          description: Valid email and explicit consent are required
        "429":
          description: Too many requests; retry later
        "503":
          description: Signup could not be saved; retry later
...
  /member/profile:
    get:
      operationId: getMemberProfile
      responses:
        "200":
          description: Member profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemberProfile" }
        "401": { description: Sign in required }
    put:
      operationId: saveMemberProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemberProfile" }
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemberProfile" }
        "400": { description: Invalid profile }
        "401": { description: Sign in required }
  /member/favourites:
    get:
      operationId: getMemberFavourites
      responses:
        "200":
          description: Saved recipes and favourite foods
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MemberFavourite" }
        "401": { description: Sign in required }
    put:
      operationId: setMemberFavourite
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FavouriteInput" }
      responses:
        "200":
          description: Updated favourites
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MemberFavourite" }
        "400": { description: Invalid favourite }
        "401": { description: Sign in required }
  /member/plans:
    get:
      operationId: getMemberPlans
      responses:
        "200":
          description: Saved daily and weekly plans
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MealPlan" }
        "401": { description: Sign in required }
    post:
      operationId: createMemberPlan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MealPlanInput" }
      responses:
        "201":
          description: Saved plan
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MealPlan" }
        "400": { description: Invalid plan }
        "401": { description: Sign in required }
  /member/plans/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    put:
      operationId: updateMemberPlan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MealPlanInput" }
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MealPlan" }
        "400": { description: Invalid plan }
        "401": { description: Sign in required }
        "404": { description: Plan not found }
    delete:
      operationId: deleteMemberPlan
      responses:
        "204": { description: Plan deleted }
        "400": { description: Invalid plan identifier }
        "401": { description: Sign in required }
        "404": { description: Plan not found }
  /member/enquiries:
    get:
      operationId: getMemberEnquiries
      responses:
        "200":
          description: Personal order enquiries, not confirmed orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MemberEnquiry" }
        "401": { description: Sign in required }
...
components:
  schemas:
    AssistantInput:
      type: object
      required: [text, requestId, useProfile]
      properties:
        text: { type: string, minLength: 1, maxLength: 2000 }
        requestId: { type: string, format: uuid }
        useProfile: { type: boolean }
    AssistantTurn:
      type: object
      required: [id, requestId, userText, useProfile, reply, status, error, createdAt]
      properties:
        id: { type: integer }
        requestId: { type: string, format: uuid }
        userText: { type: string }
        useProfile: { type: boolean }
        reply: { type: [string, "null"] }
        status: { type: string, enum: [pending, complete, failed] }
        error: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
    AssistantHistory:
      type: object
      required: [turns, dailyRemaining]
      properties:
        turns:
          type: array
          items: { $ref: "#/components/schemas/AssistantTurn" }
        dailyRemaining: { type: integer }
    ChallengeMembershipInput:
      type: object
      required: [joined]
      properties:
        joined: { type: boolean }
    ChallengeCheckinInput:
      type: object
      required: [completed]
      properties:
        completed: { type: boolean }
...
    ChallengeProgress:
      type: object
      required: [challengeSlug, isActive, startDate, endDate, checkedDates, completedDays, streak, todayDone, canCheckIn, dayNumber, isFinished]
      properties:
        challengeSlug: { type: string }
        isActive: { type: boolean }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        checkedDates:
          type: array
          items: { type: string, format: date }
        completedDays: { type: integer }
        streak: { type: integer }
        todayDone: { type: boolean }
        canCheckIn: { type: boolean }
        dayNumber: { type: integer }
        isFinished: { type: boolean }
    NewsletterSignupInput:
      type: object
      required: [email, consent]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        consent: { type: boolean, enum: [true] }
        website: { type: string, maxLength: 200, description: Anti-spam field; leave empty }
    Article:
      type: object
      required: [slug, title, category, summary, readMinutes, sections]
      properties:
        slug: { type: string }
        title: { type: string }
        category: { type: string }
        summary: { type: string }
        readMinutes: { type: integer }
        sections:
          type: array
          items:
            type: object
            required: [heading, paragraphs]
            properties:
              heading: { type: string }
              paragraphs:
                type: array
                items: { type: string }
...
    MemberProfile:
      type: object
      required: [displayName, goal, dietaryPreference, dailyCalories, dailyProteinG]
      properties:
        displayName: { type: string, maxLength: 120 }
        goal: { type: string, enum: [healthy, lose, muscle, energy, workout, office] }
        dietaryPreference: { type: string, enum: [any, vegetarian] }
        dailyCalories: { type: integer, minimum: 500, maximum: 10000 }
        dailyProteinG: { type: integer, minimum: 0, maximum: 500 }
    MemberFavourite:
      type: object
      required: [kind, slug]
      properties:
        kind: { type: string, enum: [food, recipe] }
        slug: { type: string, minLength: 1, maxLength: 150 }
    FavouriteInput:
      type: object
      required: [kind, slug, saved]
      properties:
        kind: { type: string, enum: [food, recipe] }
        slug: { type: string, minLength: 1, maxLength: 150 }
        saved: { type: boolean }
    MealEntry:
      type: object
      required: [day, meal, kind, slug, servings]
      properties:
        day: { type: integer, minimum: 0, maximum: 6 }
        meal: { type: string, enum: [breakfast, lunch, dinner, snack] }
        kind: { type: string, enum: [food, recipe] }
        slug: { type: string, minLength: 1, maxLength: 150 }
        servings: { type: number, minimum: 0.25, maximum: 20 }
    MealPlanInput:
      type: object
      required: [name, days, entries]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        days: { type: integer, enum: [1, 7] }
        entries:
          type: array
          maxItems: 200
          items: { $ref: "#/components/schemas/MealEntry" }
    NutritionTotals:
      type: object
      required: [calories, proteinG, carbsG, fatG, fiberG, costInRupees]
      properties:
        calories: { type: number }
        proteinG: { type: number }
        carbsG: { type: number }
        fatG: { type: number }
        fiberG: { type: number }
        costInRupees: { type: number }
    MealPlan:
      type: object
      required: [id, name, days, entries, totals, dailyTotals, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        days: { type: integer, enum: [1, 7] }
        entries:
          type: array
          items: { $ref: "#/components/schemas/MealEntry" }
        totals: { $ref: "#/components/schemas/NutritionTotals" }
        dailyTotals:
          type: array
          items: { $ref: "#/components/schemas/NutritionTotals" }
        updatedAt: { type: string, format: date-time }
    MemberEnquiry:
      type: object
      required: [id, kind, enquiryId, subject, detail, status, submittedAt, unreadCount]
      properties:
        id: { type: string }
        kind: { type: string, enum: [order, partnership, contact] }
        enquiryId: { type: integer, minimum: 1 }
        subject: { type: string }
        detail: { type: string }
        unreadCount: { type: integer, minimum: 0 }
        status: { type: string }
        submittedAt: { type: string, format: date-time }
        totalInRupees: { type: integer }
        itemCount: { type: integer }
        pickupLocation: { type: string }
        pickupTime: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/FoodOrderItemInput" }
...
    FoodOrderItemInput:
      type: object
      required: [foodSlug, quantity]
      properties:
        foodSlug: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 1, maximum: 20 }
...