# lib/api-zod/src/generated/types/assistantTurn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Nutrio food discovery and enquiry API
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantTurnStatus } from './assistantTurnStatus';

export interface AssistantTurn {
  id: number;
  requestId: string;
  userText: string;
  useProfile: boolean;
  /** @nullable */
  reply: string | null;
  status: AssistantTurnStatus;
  /** @nullable */
  error: string | null;
  createdAt: Date;
}