# lib/api-zod/src/generated/types/mealPlan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Nutrio food discovery and enquiry API
 * OpenAPI spec version: 0.1.0
 */
import type { MealEntry } from './mealEntry';
import type { MealPlanDays } from './mealPlanDays';
import type { NutritionTotals } from './nutritionTotals';

export interface MealPlan {
  id: number;
  name: string;
  days: MealPlanDays;
  entries: MealEntry[];
  totals: NutritionTotals;
  dailyTotals: NutritionTotals[];
  updatedAt: Date;
}